login/logout with session with functioning 404 page (CH11)
We used sessions with the express-session middleware to keep users logged
in to our app by keeping their information stored in the browser and then
sending that information to the server each time a user makes a request. We
protect pages reserved for logged in users with an authentication middleware
and conditionally display 'New Post', 'Login' and 'New User' links based on if
a user is logged in or out. We also created a 404 page for non-existing routes.